feat: 序列化转换器

+++ src/main/java/com/example/shop/model/SpuDetailImg.java
@@ -14,6 +14,6 @@ public class SpuDetailImg extends BaseEntity {
     @Id
     private Long id;
     private String img;
-    private Integer spuId;
+    private Long spuId;
     private Long index;
 }
