+++ src/main/java/com/example/shop/model/SpuImg.java
@@ -14,5 +14,5 @@ public class SpuImg extends BaseEntity {
     @Id
     private Long id;
     private String img;
-    private Integer spuId;
+    private Long spuId;
 }
